phase_4_rag/requirements.txt
google-generativeai
supabase
pytest
python-dotenv
